@@ -98,7 +98,7 @@
     "ts-jest": "^23.10.4",
     "tslint": "^5.12.0",
     "tslint-react": "^3.6.0",
-    "typescript": "2.8.3"
+    "typescript": "2.9.2"
   },
   "dependencies": {
     "@progress/kendo-data-query": "^1.5.0",
