@@ -1,8 +1,5 @@
-import {getType} from 'typesafe-actions';
-import {EmptyAction} from 'typesafe-actions/dist/types';
-import {Period, TemporalResolution} from '../../../components/dates/dateModels';
-import {SelectionInterval} from '../../../state/user-selection/userSelectionModels';
-import {Action} from '../../../types/Types';
+import {ActionType, getType} from 'typesafe-actions';
+import {Period} from '../../../components/dates/dateModels';
 import {logoutUser} from '../../auth/authActions';
 import {setMeterDetailsTimePeriod} from './meterDetailActions';
 import {MeterDetailState} from './meterDetailModels';
@@ -11,15 +8,12 @@ const initialState: MeterDetailState = {
   timePeriod: {period: Period.latest},
 };
 
-type ActionTypes = Action<TemporalResolution | SelectionInterval> | EmptyAction<string>;
+type ActionTypes = ActionType<typeof setMeterDetailsTimePeriod | typeof logoutUser>;
 
-export const meterDetail = (
-  state: MeterDetailState = initialState,
-  action: ActionTypes
-): MeterDetailState => {
+export const meterDetail = (state: MeterDetailState = initialState, action: ActionTypes): MeterDetailState => {
   switch (action.type) {
     case getType(setMeterDetailsTimePeriod):
-      return {...state, timePeriod: {...(action as Action<SelectionInterval>).payload}};
+      return {...state, timePeriod: action.payload};
     case getType(logoutUser):
       return initialState;
     default:
